apps/tts-service/requirements.txt
flask
numpy==1.26.4
torch
torchaudio
scipy
python-dotenv